Country dictionary

Table of countries, areas, and territories with their official names, M49 and ISO codes, and regional groupings.

Usage

countrynames

Format

An object of class tbl_df (inherits from tbl, data.frame) with 257 rows and 21 columns.

Details

This is the UN's Standard Country or Area Codes for Statistical Use, also known as the M49 standard (https://unstats.un.org/unsd/methodology/m49), with the following revisions:

"State of Palestine" is replaced with "Occupied Palestinian Territory" in accordance with the IOM House Style Manual (May 2023).

"Kosovo (SCR 1244)" is added with ISO codes "XK" and "XKX". References to Kosovo shall be understood to be in the context of United Nations Security Council resolution 1244 (1999).

"Channel Islands" is added with ISO codes "XC" and "XCX". The Channel Islands comprise the British crown dependencies of Guernsey and Jersey, which are present in the table. However, some datasets (such as UN DESA's international migrant stocks) combine Guernsey and Jersey under the Channel Islands, so the separate entry accounts for such cases.

"Abyei" is added with ISO-3 code "AB9". This follows the designations of the IDMC, whose data separately identify this disputed area on the border between Sudan and South Sudan.

"Taiwan Province of China, China" is added with ISO-3 code "TWN".

"Stateless" is added with ISO-3 code "STL".

"Unknown" is added with ISO-3 code "OOO".

m49

double This is the UN's Standard Country or Area Codes for Statistical Use, also known as the M49 standard (https://unstats.un.org/unsd/methodology/m49)

iso2

character ISO 2 Characters

iso3

character ISO 3 Characters

name

character COLUMN_DESCRIPTION

name_sort

character COLUMN_DESCRIPTION

name_text

character COLUMN_DESCRIPTION

with_the

double COLUMN_DESCRIPTION

region_code

double COLUMN_DESCRIPTION

region

character COLUMN_DESCRIPTION

subregion_code

double COLUMN_DESCRIPTION

subregion

character COLUMN_DESCRIPTION

intregion_code

double COLUMN_DESCRIPTION

intregion

character COLUMN_DESCRIPTION

rdh

character COLUMN_DESCRIPTION

ldc

double COLUMN_DESCRIPTION

lldc

double COLUMN_DESCRIPTION

sids

double COLUMN_DESCRIPTION

eu

double membership in the European Union

schengen

double Membership in the Schengen Area

balkans

double the Balkans

iom.region

double IOM Regions as defined in the appeal as displayed here: https://www.iom.int/sites/g/files/tmzbdl486/files/appeals/iom-global-appeal-2024_final.pdf#page=21